+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Workspace demo documentation
 
+### Fixed
+
+- Use `product.link` instead of `slugifyUrl`.
+
 ## [0.5.0] - 2020-01-13
 
 ### Changed